+++ models/__init__.py
@@ -4,3 +4,4 @@ db = SQLAlchemy()
 
 from .message import Message
 from .device import Device
+from .api_key import ApiKey